update:可読性向上のためクラスをclass.rbに分割
